requirements.txt
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonobject==0.9.10
MarkupSafe==2.0.1
pymongo==3.12.0
pytz==2021.3
PyYAML==6.0
scapy==2.4.5
six==1.16.0
Werkzeug==2.0.2
